--- a/deeplabcut/pose_estimation_pytorch/modelzoo/inference.py
+++ b/deeplabcut/pose_estimation_pytorch/modelzoo/inference.py
@@ -27,7 +27,7 @@ from deeplabcut.pose_estimation_pytorch.modelzoo.utils import (
 )
 from deeplabcut.utils.make_labeled_video import create_video
 
-
+    
 class NumpyEncoder(json.JSONEncoder):
     """Special json encoder for numpy types"""
 
@@ -120,7 +120,7 @@ def _video_inference_superanimal(
 
     for video_path in video_paths:
         print(f"Processing video {video_path}")
-
+        
         dlc_scorer = get_super_animal_scorer(
             superanimal_name, model_snapshot_path, detector_snapshot_path
         )
@@ -171,6 +171,7 @@ def _video_inference_superanimal(
 
         superanimal_colormaps = get_superanimal_colormaps()
         colormap = superanimal_colormaps[superanimal_name]
+
         create_video(
             video_path,
             output_h5,
@@ -182,7 +183,7 @@ def _video_inference_superanimal(
             plot_bboxes=plot_bboxes,
             bboxes_list=bboxes_list,
             bboxes_pcutoff=bboxes_pcutoff,
-        )
+        )        
         print(f"Video with predictions was saved as {output_path}")
 
     return results