--- a/requirements.txt
+++ b/requirements.txt
@@ -24,7 +24,6 @@ scikit-image>=0.17
 scikit-learn>=1.0
 scipy>=1.9
 statsmodels>=0.11
-tables==3.8.0
-torch==2.6.0
+torch==2.7.1
 torchvision
 tqdm